requirements.txt
kafka_python==2.1.4
psycopg2==2.9.10
python-dotenv==1.1.0
